Split out XeTeX classes (again)
Looking at this in detail, it is clear that there is a difference
between setting up the other TeX data and setting up inter character
classes. The other data map well between Unicode and TeX and there is
no question of needing further macro-level support to make use of
the definitions here. On the other hand, for the inter character
classes it is clear that several 'raw' classes are needed for e.g.
CJK but these should be presented to the user in various different
guises.